--- App.css
@@ -1,6 +1,0 @@
-.App {
-  min-height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
