+++ src/app/layout.tsx
@@ -5,8 +5,8 @@ import "./globals.css";
 const manrope = Manrope({subsets:["latin"]});
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Akira - Web3 Advertising Platform",
+  description: "Decentralized ads for Web3 growth",
 };
 
 export default function RootLayout({
