+++ src/lib/solana/idl.json
@@ -97,4 +97,4 @@
       }
     }
   ]
-}
+}
