+++ src/lib/solana/provider.tsx
@@ -1,0 +1,44 @@
+import { WalletAdapterNetwork } from '@solana/wallet-adapter-base';
+import { ConnectionProvider, WalletProvider } from '@solana/wallet-adapter-react';
+import { WalletModalProvider } from '@solana/wallet-adapter-react-ui';
+import { UnsafeBurnerWalletAdapter } from '@solana/wallet-adapter-wallets';
+import { clusterApiUrl } from '@solana/web3.js';
+import React, { FC, ReactNode, useMemo } from 'react';
+
+// Required for wallet adapter UI
+import '@solana/wallet-adapter-react-ui/styles.css';
+
+export const SolanaProvider: FC<{ children: ReactNode }> = ({ children }) => {
+    // The network can be set to 'devnet', 'testnet', or 'mainnet-beta'.
+    const network = WalletAdapterNetwork.Devnet;
+
+    // You can also provide a custom RPC endpoint.
+    const endpoint = useMemo(() => clusterApiUrl(network), [network]);
+
+    const wallets = useMemo(
+        () => [
+            /**
+             * Wallets that implement either of these standards will be available automatically.
+             *
+             *   - Solana Mobile Stack Mobile Wallet Adapter Protocol
+             *     (https://github.com/solana-mobile/mobile-wallet-adapter)
+             *   - Solana Wallet Standard
+             *     (https://github.com/solana-labs/wallet-standard)
+             *
+             * If you wish to support a wallet that supports neither of those standards,
+             * instantiate its legacy wallet adapter here.
+             */
+            new UnsafeBurnerWalletAdapter(),
+        ],
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+        [network]
+    );
+
+    return (
+        <ConnectionProvider endpoint={endpoint}>
+            <WalletProvider wallets={wallets} autoConnect>
+                <WalletModalProvider>{children}</WalletModalProvider>
+            </WalletProvider>
+        </ConnectionProvider>
+    );
+};
