+++ src/types/global.d.ts
@@ -1,0 +1,7 @@
+import { Buffer } from 'buffer';
+
+declare global {
+  interface Window {
+    Buffer: typeof Buffer;
+  }
+}
